Cuenta

+++ backend/scr/DigitalArs/Dtos/CreateCuentaDtos.cs
@@ -4,7 +4,7 @@ namespace DigitalArs.Dtos
     {
         public required int ID_CUENTA { get; set; }
         public required int ID_USUARIO { get; set; }
-        public required float SALDO { get; set; }
+        public required Double SALDO { get; set; }
         public required  string ALIAS { get; set; }
         public required string  CBU { get; set; }
     }
